src/requirements.txt
python-dotenv
controlhub